@@ -28,11 +28,15 @@
     "@react-native-firebase/firestore": "^15.2.0",
     "@react-native-firebase/messaging": "^15.1.1",
     "@react-native-firebase/remote-config": "^15.1.1",
+    "@react-native-picker/picker": "^2.4.2",
     "@react-navigation/bottom-tabs": "^6.3.2",
     "@react-navigation/native": "^6.0.11",
     "@react-navigation/stack": "^6.2.2",
     "@redux-saga/testing-utils": "^1.1.3",
+    "airbnb": "^0.0.2",
     "buffer": "^6.0.3",
+    "cross-fetch": "^3.1.5",
+    "eslint-config-airbnb": "^19.0.4",
     "jwt-decode": "^3.1.2",
     "luxon": "^1.24.1",
     "react": "^18.2.0",
@@ -75,16 +79,17 @@
     "@types/react-redux": "^7.1.24",
     "@types/redux-logger": "^3.0.9",
     "@types/redux-saga": "^0.10.5",
-    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/eslint-plugin": "^5.30.7",
     "@typescript-eslint/parser": "^5.30.6",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^28.1.3",
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-babel": "^5.3.0",
     "eslint-plugin-fp": "^2.3.0",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.0.0",
@@ -96,10 +101,11 @@
     "metro-config": "^0.71.3",
     "metro-react-native-babel-preset": "^0.71.3",
     "prettier": "^2.7.1",
+    "react-native-flipper": "^0.155.0",
     "react-native-svg-transformer": "^1.0.0",
     "react-native-typescript-transformer": "^1.2.13",
     "react-test-renderer": "^18.2.0",
-    "redux-devtools-extension": "^2.13.9R",
+    "redux-flipper": "^2.0.2",
     "redux-logger": "^3.0.6",
     "ts-jest": "^28.0.6",
     "typescript": "^4.7.4"
