@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import {
   StyleProp,
   StyleSheet,
@@ -7,12 +7,9 @@ import {
   TextStyle,
   View,
   ViewStyle,
+  KeyboardTypeOptions,
 } from "react-native";
-import Colors from "../constants/Colors";
-
-interface State {
-  isFocused: boolean;
-}
+import { Colors } from "../constants/Colors";
 
 interface Props {
   value: string;
@@ -20,7 +17,7 @@ interface Props {
   hideLabelWhenFocused?: boolean;
   required?: boolean;
   maxLength?: number;
-  keyboardType?: string;
+  keyboardType?: KeyboardTypeOptions;
   onChangeText?: (value: string) => void;
   onFocus?: () => void;
   defaultValue?: string;
@@ -30,92 +27,82 @@ interface Props {
   secureTextEntry?: boolean;
 }
 
-export default class InputWithLabel extends Component<Props, State> {
-  state = {
-    isFocused: false,
+export const InputWithLabel = (props: Props) => {
+  const [isFocused, setIsFocused] = useState<boolean>(false);
+
+  const onFocus = () => setIsFocused(true);
+
+  const onBlur = () => setIsFocused(false);
+
+  const {
+    label,
+    value,
+    hideLabelWhenFocused,
+    required,
+    maxLength,
+    style,
+    error,
+    placeholder,
+    secureTextEntry,
+  } = props;
+
+  const keyboardType = props.keyboardType ? props.keyboardType : "default";
+
+  const isEmpty = !value || value.length === 0;
+
+  const isSmallLabel = isFocused || !isEmpty;
+
+  const labelStyle: TextStyle = {
+    position: "absolute",
+    left: 10,
+    color: Colors.black,
+    top: isSmallLabel ? 4 : 14,
+    fontSize: isSmallLabel ? 12 : 16,
   };
 
-  onFocus = () => this.setState({ isFocused: true });
+  const showRequired = required && !value;
 
-  onBlur = () => this.setState({ isFocused: false });
+  return (
+    <View style={style}>
+      {hideLabelWhenFocused ? (
+        <Text style={styles.outsideLabel}>{label}</Text>
+      ) : null}
+      {!hideLabelWhenFocused ? <Text style={labelStyle}>{label}</Text> : null}
 
-  render() {
-    const {
-      label,
-      value,
-      hideLabelWhenFocused,
-      required,
-      maxLength,
-      style,
-      error,
-      placeholder,
-      secureTextEntry,
-      ...props
-    } = this.props;
-    const { isFocused } = this.state;
+      {showRequired ? (
+        <Text style={styles.requiredLabelStyle}>Required</Text>
+      ) : null}
 
-    const keyboardType = this.props.keyboardType
-      ? this.props.keyboardType
-      : "default";
+      {error ? <Text style={styles.requiredLabelStyle}>{error}</Text> : null}
 
-    const isEmpty = !value || value.length === 0;
+      {maxLength ? (
+        <Text style={styles.maxLength}>
+          {value.length}/{maxLength}
+        </Text>
+      ) : null}
 
-    const isSmallLabel = isFocused || !isEmpty;
-
-    const labelStyle: TextStyle = {
-      position: "absolute",
-      left: 10,
-      color: Colors.black,
-      top: isSmallLabel ? 4 : 14,
-      fontSize: isSmallLabel ? 12 : 16,
-    };
-
-    const showRequired = required && !value;
-
-    return (
-      <View style={style}>
-        {hideLabelWhenFocused ? (
-          <Text style={styles.outsideLabel}>{label}</Text>
-        ) : null}
-        {!hideLabelWhenFocused ? <Text style={labelStyle}>{label}</Text> : null}
-
-        {showRequired ? (
-          <Text style={styles.requiredLabelStyle}>Required</Text>
-        ) : null}
-
-        {error ? <Text style={styles.requiredLabelStyle}>{error}</Text> : null}
-
-        {maxLength ? (
-          <Text style={styles.maxLength}>
-            {value.length}/{maxLength}
-          </Text>
-        ) : null}
-
-        <TextInput
-          {...props}
-          defaultValue={value}
-          style={[
-            styles.textInput,
-            isFocused ? styles.focused : null,
-            hideLabelWhenFocused ? styles.textInputWithNoLabel : null,
-            maxLength ? styles.textInputWithCounter : null,
-            showRequired ? { borderBottomColor: Colors.required } : null,
-          ]}
-          onFocus={this.onFocus}
-          onBlur={this.onBlur}
-          // eslint-disable-line react/jsx-tag-spacing
-          // @ts-ignore
-          keyboardType={keyboardType}
-          maxLength={maxLength}
-          autoCapitalize="none"
-          underlineColorAndroid="transparent"
-          secureTextEntry={secureTextEntry}
-          placeholder={placeholder && !label ? placeholder : ""}
-        />
-      </View>
-    );
-  }
-}
+      <TextInput
+        {...props}
+        defaultValue={value}
+        style={[
+          styles.textInput,
+          isFocused ? styles.focused : null,
+          hideLabelWhenFocused ? styles.textInputWithNoLabel : null,
+          maxLength ? styles.textInputWithCounter : null,
+          showRequired ? { borderBottomColor: Colors.required } : null,
+        ]}
+        onFocus={onFocus}
+        onBlur={onBlur}
+        keyboardType={keyboardType}
+        maxLength={maxLength}
+        autoCapitalize="none"
+        underlineColorAndroid="transparent"
+        secureTextEntry={secureTextEntry}
+        placeholder={placeholder && !label ? placeholder : ""}
+      />
+    </View>
+  );
+};
 
 const styles = StyleSheet.create({
   textInput: {
