@@ -1,15 +1,14 @@
-import React, { FC } from "react";
+import React from "react";
 import { StyleSheet, View } from "react-native";
 
-import Colors from "../constants/Colors";
-
 import Icon from "react-native-vector-icons/Ionicons";
+import { Colors } from "../constants/Colors";
 
 interface Props {
   size?: number;
 }
 
-const ImagePlaceholder: FC<Props> = (props: Props) => {
+export const ImagePlaceholder = (props: Props) => {
   return (
     <View style={styles.container}>
       <Icon size={props.size || 35} name="image-outline" style={styles.icon} />
@@ -33,5 +32,3 @@ const styles = StyleSheet.create({
     color: Colors.grey,
   },
 });
-
-export default ImagePlaceholder;
