@@ -1,14 +1,14 @@
-import React, { FC } from "react";
+import React from "react";
 import { StyleSheet, Text, TouchableOpacity } from "react-native";
 
-import Colors from "../constants/Colors";
+import { Colors } from "../constants/Colors";
 
 type Props = {
   text: string;
   onPress: () => void;
 };
 
-const NavigationLink: FC<Props> = ({ text, onPress }) => (
+export const NavigationLink = ({ text, onPress }: Props) => (
   <TouchableOpacity style={styles.linkContainer} onPress={onPress}>
     <Text style={styles.linkText}>{text}</Text>
   </TouchableOpacity>
@@ -24,5 +24,3 @@ const styles = StyleSheet.create({
     color: Colors.gold,
   },
 });
-
-export default NavigationLink;
