@@ -1,18 +1,18 @@
-export { default as BodyText } from "./BodyText";
-export { default as Card } from "./Card";
-export { default as Carousel } from "./Carousel";
-export { default as Button } from "./Button";
-export { default as HyperLink } from "./HyperLink";
-export { default as InputWithLabel } from "./InputWithLabel";
-export { default as ButtonWithIcon } from "./ButtonWithIcon";
-export { default as NavigationLink } from "./NavigationLink";
-export { default as PhoneNumberPrefix } from "./PhoneNumberPrefix";
-export { default as Dropdown } from "./Dropdown";
-export { default as OptionSwitch } from "./OptionSwitch";
-export { default as TextAreaWithLabel } from "./TextAreaWithLabel";
-export { default as Screen } from "./Screen";
-export { default as Divider } from "./Divider";
-export { default as GenericModal } from "./GenericModal";
-export { default as Title } from "./Title";
-export { default as ImagePlaceholder } from "./ImagePlaceholder";
-// export { default as WebViewInModal } from './WebViewInModal'
+export { BodyText } from "./BodyText";
+export { Card } from "./Card";
+export { Carousel } from "./Carousel";
+export { Button } from "./Button";
+export { HyperLink } from "./HyperLink";
+export { InputWithLabel } from "./InputWithLabel";
+export { ButtonWithIcon } from "./ButtonWithIcon";
+export { NavigationLink } from "./NavigationLink";
+export { PhoneNumberPrefix } from "./PhoneNumberPrefix";
+export { Dropdown } from "./Dropdown";
+export { OptionSwitch } from "./OptionSwitch";
+export { TextAreaWithLabel } from "./TextAreaWithLabel";
+export { Screen } from "./Screen";
+export { Divider } from "./Divider";
+export { GenericModal } from "./GenericModal";
+export { Title } from "./Title";
+export { ImagePlaceholder } from "./ImagePlaceholder";
+export { Loader } from "./Loader";
