@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { useNavigation } from "@react-navigation/native";
+import { StackNavigationProp } from "@react-navigation/stack";
 import { Alert, View, StyleSheet } from "react-native";
 import * as Progress from "react-native-progress";
 
@@ -11,7 +12,7 @@ import {
   Screen,
   Title,
 } from "../components";
-import Colors from "../constants/Colors";
+import { Colors } from "../constants/Colors";
 
 import { authActions } from "../store/actions/auth";
 import {
@@ -20,12 +21,29 @@ import {
 } from "../store/selectors";
 
 import { ScreenNames } from "../navigation/ScreenNames";
+import { NavigatorStackParamList } from "../navigation/AppNavigator";
 
 import { StringValues } from "../constants/StringValues";
 
+type RegistrationScreenProp = StackNavigationProp<
+  NavigatorStackParamList,
+  typeof ScreenNames.REGISTRATION
+>;
+
+const styles = StyleSheet.create({
+  titleContainer: {
+    paddingTop: 100,
+    paddingBottom: 12,
+  },
+  loader: {
+    paddingVertical: 24,
+    alignSelf: "center",
+  },
+});
+
 export const RegistrationScreen = () => {
   const dispatch = useDispatch();
-  const navigation = useNavigation();
+  const navigation = useNavigation<RegistrationScreenProp>();
 
   const [firstName, setFirstName] = useState<string>("");
   const [lastName, setLastName] = useState<string>("");
@@ -34,7 +52,7 @@ export const RegistrationScreen = () => {
   const [confirmPassword, setConfirmPassword] = useState<string>("");
 
   const isRegistering: boolean = useSelector(getRegistrationStatus);
-  const error: string | null = useSelector(getRegistrationError);
+  const error: boolean | null = useSelector(getRegistrationError);
 
   useEffect(() => {
     if (isRegistering === false && !error) {
@@ -77,32 +95,32 @@ export const RegistrationScreen = () => {
       </View>
 
       <InputWithLabel
-        hideLabelWhenFocused={true}
+        hideLabelWhenFocused
         value={firstName}
         onChangeText={value => setFirstName(value)}
         label={StringValues.firstName}
       />
       <InputWithLabel
-        hideLabelWhenFocused={true}
+        hideLabelWhenFocused
         value={lastName}
         onChangeText={value => setLastName(value)}
         label={StringValues.lastName}
       />
       <InputWithLabel
-        hideLabelWhenFocused={true}
+        hideLabelWhenFocused
         value={email}
         onChangeText={value => setEmail(value)}
         label={StringValues.email}
       />
       <InputWithLabel
-        hideLabelWhenFocused={true}
+        hideLabelWhenFocused
         value={password}
         onChangeText={value => setPassword(value)}
         label={StringValues.password}
         secureTextEntry
       />
       <InputWithLabel
-        hideLabelWhenFocused={true}
+        hideLabelWhenFocused
         value={confirmPassword}
         onChangeText={value => setConfirmPassword(value)}
         label={StringValues.confirmPassword}
@@ -123,7 +141,7 @@ export const RegistrationScreen = () => {
       {isRegistering ? (
         <Progress.Circle
           size={24}
-          indeterminate={true}
+          indeterminate
           color={Colors.gold}
           style={styles.loader}
         />
@@ -131,14 +149,3 @@ export const RegistrationScreen = () => {
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  titleContainer: {
-    paddingTop: 100,
-    paddingBottom: 12,
-  },
-  loader: {
-    paddingVertical: 24,
-    alignSelf: "center",
-  },
-});
